requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
gradio
requests
pydantic
rich

